Changes to reduce startup time of Micronaut App
We observed that using micronaut-liquibase has blown up the start up time of Micronaut App, even if there are no changes to be made to Database. I have taken up the code from micronaut liquibase library and did some optimizations on it to reduce the startup time of the application, when there are no changes to be made to database. This is because the "performUpdate" method is called irrespective of whether the changes are really required to be made to db. I have created another method "performUpdateIfNeeded()" that checks whether there are unrun changesets or not. If there are no unrun changesets, it won't call "performUpdate" and "liquibase.update" at all.